more CodeFactor and Coverity updates

+++ src/eve-server/inventory/InventoryDB.cpp
@@ -182,7 +182,7 @@ bool InventoryDB::GetItemContents(OwnerData &od, std::vector<uint32> &into) {
         return false;
     }
 
-    _log(DATABASE__RESULTS, "GetItemContents: '%s' returned %u items", query.str().c_str(), res.GetRowCount());
+    _log(DATABASE__RESULTS, "GetItemContents: '%s' returned %lu items", query.str().c_str(), res.GetRowCount());
     DBResultRow row;
     while( res.GetRow( row ) )
         into.push_back( row.GetUInt( 0 ) );
@@ -207,7 +207,7 @@ bool InventoryDB::GetItemContents(uint32 itemID, EVEItemFlags flag, std::vector<
         return false;
     }
 
-    _log(DATABASE__RESULTS, "GetItemContents for item %u returned %u items", itemID, res.GetRowCount());
+    _log(DATABASE__RESULTS, "GetItemContents for item %u returned %lu items", itemID, res.GetRowCount());
     DBResultRow row;
     while( res.GetRow( row ) )
         into.push_back( row.GetUInt( 0 ) );
@@ -229,11 +229,11 @@ bool InventoryDB::GetItemContents(uint32 itemID, EVEItemFlags flag, uint32 owner
         "  AND ownerID=%u",
         itemID, (int)flag, ownerID ) )
     {
-        codelog(DATABASE__ERROR, "Error in GetItemContents query for item %u with flag %u: %s", itemID, (int)flag, res.error.c_str());
+        codelog(DATABASE__ERROR, "Error in GetItemContents query for item %u with flag %i: %s", itemID, (int)flag, res.error.c_str());
         return false;
     }
 
-    _log(DATABASE__RESULTS, "GetItemContents for item %u with flag %u returned %u items", itemID, flag, res.GetRowCount());
+    _log(DATABASE__RESULTS, "GetItemContents for item %u with flag %i returned %lu items", itemID, (int)flag, res.GetRowCount());
     DBResultRow row;
     while( res.GetRow( row ) )
         into.push_back( row.GetUInt( 0 ) );
